requirements.txt
chromadb==1.3.5
google-generativeai==0.8.5
pandas==2.3.3
pillow==12.0.0
PyYAML==6.0.3
streamlit==1.51.0
watchdog==6.0.0